Add PostgreSQL code to migrations

+++ database/migrations/2021_05_17_125729_create_procedure.php
@@ -1,8 +1,6 @@
 <?php
 
 use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
-use Illuminate\Support\Facades\Schema;
 
 class CreateProcedure extends Migration
 {
@@ -13,33 +11,67 @@ class CreateProcedure extends Migration
      */
     public function up()
     {
-        $statement = "CREATE PROCEDURE MakeOrder @quantity INT, @item_id INT, @result INT OUTPUT
-                        AS
-                        DECLARE @device_id INT, @order_id INT, @total_quantity INT, @new_quantity INT, @sum FLOAT
-                        BEGIN
-                            SET @device_id = (SELECT d.id FROM items i inner join devices d on i.device_id = d.id WHERE i.item_id = @item_id)
-                            SET @order_id = (SELECT order_id FROM items WHERE item_id = @item_id)
-                            SET @total_quantity = (SELECT total_quantity FROM devices WHERE id = @device_id)
-                            
-                            IF @quantity <= @total_quantity
+        // MS SQL Server code
+        if (DB::connection() instanceof \Illuminate\Database\SqlServerConnection) {
+            $statement = "CREATE PROCEDURE MakeOrder @quantity INT, @item_id INT, @result INT OUTPUT
+                            AS
+                            DECLARE @device_id INT, @order_id INT, @total_quantity INT, @new_quantity INT, @sum FLOAT
                             BEGIN
-                                BEGIN TRANSACTION
-                                    SET @sum = dbo.getSum(@quantity, @device_id)
-                                    SET @new_quantity = @total_quantity - @quantity
-                                    UPDATE items SET total_sum = @sum WHERE item_id = @item_id
-                                    UPDATE devices SET total_quantity = @new_quantity WHERE id = @device_id
-                                    SET @result = 1
-                                COMMIT TRANSACTION
-                            END
-                            ELSE
-                            BEGIN
-                                BEGIN TRANSACTION
-                                    DELETE FROM orders WHERE id = @order_id
-                                COMMIT TRANSACTION
-                            END
-                        END";
+                                SET @device_id = (SELECT d.id FROM items i inner join devices d on i.device_id = d.id WHERE i.item_id = @item_id)
+                                SET @order_id = (SELECT order_id FROM items WHERE item_id = @item_id)
+                                SET @total_quantity = (SELECT total_quantity FROM devices WHERE id = @device_id)
+                                
+                                IF @quantity <= @total_quantity
+                                BEGIN
+                                    BEGIN TRANSACTION
+                                        SET @sum = dbo.getSum(@quantity, @device_id)
+                                        SET @new_quantity = @total_quantity - @quantity
+                                        UPDATE items SET total_sum = @sum WHERE item_id = @item_id
+                                        UPDATE devices SET total_quantity = @new_quantity WHERE id = @device_id
+                                        SET @result = 1
+                                    COMMIT TRANSACTION
+                                END
+                                ELSE
+                                BEGIN
+                                    BEGIN TRANSACTION
+                                        DELETE FROM orders WHERE id = @order_id
+                                    COMMIT TRANSACTION
+                                END
+                            END";
 
-        DB::statement($statement);
+            DB::statement($statement);
+        }
+
+        // PostgreSQL code
+        if (DB::connection() instanceof \Illuminate\Database\PostgresConnection) {
+            $statement = "create or replace procedure make_order(quantity int, i_id int)
+                            language plpgsql    
+                            as
+                            $$
+                            declare
+                                device_id int; 
+                                order_id int; 
+                                total_quantity int; 
+                                new_quantity int;
+                                sum real;
+                            begin
+                                device_id := (SELECT d.id FROM items i inner join devices d on i.device_id = d.id WHERE i.item_id = i_id);
+                                order_id := (SELECT order_id FROM items WHERE item_id = i_id);
+                                total_quantity := (SELECT total_quantity FROM devices WHERE id = device_id);
+                            
+                                IF quantity <= total_quantity
+                                THEN
+                                    sum := get_sum(quantity, device_id);
+                                    new_quantity := total_quantity - quantity;
+                                    UPDATE items SET total_sum = sum WHERE item_id = i_id;
+                                    UPDATE devices SET total_quantity = new_quantity WHERE id = device_id;
+                                ELSE
+                                    DELETE FROM orders WHERE id = order_id;
+                                END IF;
+                            end;$$";
+
+            DB::statement($statement);
+        }
     }
 
     /**
@@ -49,7 +81,16 @@ class CreateProcedure extends Migration
      */
     public function down()
     {
-        $statement = "DROP PROCEDURE MakeOrder";
-        DB::statement($statement);
+        // MS SQL Server code
+        if (DB::connection() instanceof \Illuminate\Database\SqlServerConnection) {
+            $statement = "DROP PROCEDURE MakeOrder";
+            DB::statement($statement);
+        }
+
+        // PostgreSQL code
+        if (DB::connection() instanceof \Illuminate\Database\PostgresConnection) {
+            $statement = "DROP PROCEDURE make_order;";
+            DB::statement($statement);
+        }
     }
 }
